requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
Werkzeug==0.10.4
itsdangerous==0.24
wsgiref==0.1.2
